Move xmlToArray function to Misc
